==> Payroll Final Project/Mac Payroll/Payroll System/1 Tax Calculator.ipynb <==
import pandas as pd
import csv

user_name = input("User Name: ")
directory1 = "/Users/"
directory2 = "/Desktop/Payroll System"
directory = directory1 + user_name + directory2

Employee_Info_W_Tax_df = pd.read_csv(directory + '/1 Information CSVs/Employee-Information.csv')
Philhealth_df = pd.read_csv(directory + '/2 Tax Information/Philhealth.csv')
Pagibig_df = pd.read_csv(directory + '/2 Tax Information/Pagibig.csv')
SSS_df = pd.read_csv(directory + '/2 Tax Information/SSS.csv')

Employee_Info_W_Tax_df.loc[:,'Rate Per Month'] = Employee_Info_W_Tax_df.loc[:,'Rate Per Month'].apply(pd.to_numeric)
Employee_Info_W_Tax_df.dtypes

def IncomeTax(x):
    if x <= 250000/12:
        IncomeTax = 0
    elif x > 250000/12 and x <= 400000/12:
        IncomeTax = (x-250000/12)*0.15
    elif x > 400000/12 and x <= 800000:
        IncomeTax = 22500/12 + (x-400000/12)*0.20
    elif x > 800000/12 and x <= 2000000/12:
        IncomeTax = 102500/12 + (x-800000/12)*0.25
    elif x > 2000000/12 and x <= 8000000/12:
        IncomeTax = 402500/12 + (x-2000000/12)*0.30
    else:
        IncomeTax = 2202500/12 + (x-8000000/12)*0.35
    return IncomeTax

def Philhealth(x):
    if x <= 10000:
        Philhealth = 137.5
    elif x > 10000 and x <= 39999.99:
        Philhealth = x*0.0275/2
    else:
        Philhealth = 550
    return Philhealth

def Pagibig(x):
    if x <= 1500:
        Pagibig = x*0.001
    else:
        Pagibig = x*0.002
    return Pagibig

# Dictionary Version for csv
import csv
with open(directory + '/2 Tax Information/SSS.csv',newline="") as csv_file:
    SSS = csv.DictReader(csv_file)
    SSS_csv = []
    for i in SSS:
        Bottom_value_of_Range = i['Bottom value of Range']
        Highest_value_of_range = i['Highest value of range']
        Monthly_Salary_Credit = i['Monthly Salary Credit']
        Employee_Contribution = i['Employee Contribution']
        Employer_Contribution = i['Employer Contribution']
        SSS_csv.append([Bottom_value_of_Range,Highest_value_of_range,Monthly_Salary_Credit,Employee_Contribution,Employer_Contribution])

SSS_csv.pop()

#SSS without "AND ABOVE"
SSS_csv

import csv
with open(directory + '/2 Tax Information/SSS.csv',newline="") as csv_file:
    SSS = csv.DictReader(csv_file)
    
    def SSS_Employee_Contribution(x):
        Contribution = 0
        if x > 19750:
            Contribution = 800
        for i in SSS_csv:
            if x > float(i[0]) and x < float(i[1]):
                Contribution = int(i[3])
        return Contribution
    
    def SSS_Employer_Contribution(x):
        Contribution = 0
        for i in SSS_csv:
            if x > 19750:
                Contribution = 1630
            if x > float(i[0]) and x < float(i[1]):
                Contribution = int(i[4])
        return Contribution

    #Monthly Taxes
    Employee_Info_W_Tax_df.loc[:,'Income Tax'] = Employee_Info_W_Tax_df.loc[:,'Rate Per Month'].apply(lambda x: IncomeTax(x)).round(2)
    Employee_Info_W_Tax_df.loc[:,'Philhealth'] = Employee_Info_W_Tax_df.loc[:,'Rate Per Month'].apply(lambda x: Philhealth(x)).round(2)
    Employee_Info_W_Tax_df.loc[:,'Pagibig'] = Employee_Info_W_Tax_df.loc[:,'Rate Per Month'].apply(lambda x: Pagibig(x)).round(2)
    Employee_Info_W_Tax_df.loc[:,'SSS Employee Contribution'] = Employee_Info_W_Tax_df.loc[:,'Rate Per Month'].apply(lambda x: SSS_Employee_Contribution(x))
    Employee_Info_W_Tax_df.loc[:,'SSS Employer Contribution'] = Employee_Info_W_Tax_df.loc[:,'Rate Per Month'].apply(lambda x: SSS_Employer_Contribution(x))
    
    #Daily Taxes
    Employee_Info_W_Tax_df['Income Tax Day'] = (Employee_Info_W_Tax_df.loc[:,'Income Tax']*1/30).round(2)
    Employee_Info_W_Tax_df['Philhealth Day'] = (Employee_Info_W_Tax_df.loc[:,'Philhealth']*1/30).round(2)
    Employee_Info_W_Tax_df['Pagibig Day'] = (Employee_Info_W_Tax_df.loc[:,'Pagibig']*1/30).round(2)
    Employee_Info_W_Tax_df['SSS Employee Contribution Day'] = (Employee_Info_W_Tax_df.loc[:,'SSS Employee Contribution']/30).round(2)
    Employee_Info_W_Tax_df['SSS Employer Contribution Day'] = (Employee_Info_W_Tax_df.loc[:,'SSS Employer Contribution']/30).round(2)


Employee_Info_W_Tax_df

Output_csv = Employee_Info_W_Tax_df.to_csv(directory + '/1 Information CSVs/Employee-Information-With-Tax.csv')
