==> Payroll Final Project/Mac Payroll/Payroll System/5 Pay Slip Generator.ipynb <==
import pandas as pd
import html

user_name = input("User Name: ")
directory1 = "/Users/"
directory2 = "/Desktop/Payroll System"
directory = directory1 + user_name + directory2

folder_name = str(input("Name of Folder: "))
one_week_statistic_name = input("Write the name of the one week statistic: ")
directory_folder_statname = directory + "/" + folder_name + "/" + one_week_statistic_name + ".csv"

one_week_statistic_df = pd.read_csv(directory_folder_statname)
OWStransposed = one_week_statistic_df.transpose()
ows_dict = OWStransposed.to_dict('series')

OWStransposed

main_html = """

<html>
    <head>
        <title>Pay Slip for {0} {1} </title>
    </head>
    <body>
    
<p><strong>Employee Payroll</strong>
<br />-----------------------------------------------------------------------------------------<br />
Name: {0} {1} <br />
ID Number: {2} <br />
Date Started: {3}<br />
Rate Per Day: {4}<br />
Rate Per Month: {5}</p>

<p>Total Work Hours: {6}<br />
Overtime Hours: {7}<br />
Late Hours: {8}
<br />------------------------------------------------------------------------------------------<br />
Basic Salary:............................................{9}<br />
Overtime Pay:..........................................{10}<br />
Late Deduction:........................................{11}<br />
Total Gross Salary:...................................{12} <br />
Total Debt:................................................{13}<br />
Deductibles:.............................................{14}<br />
Bonuses:..................................................{15}</p>

<p><strong>Standing Salary:</strong>.........................................<strong>{16}<br />
</strong>-------------------------------------------------------------------------------------------<br />
Income Tax:..............................................{17}<br />
Philhealth:.................................................{18}<br />
Pagibig:.....................................................{19}<br />
SSS Employee Contribution:....................{20}<br />
SSS Employer Contribution:.....................{21}</p>
<p><strong>Salary of the Week:.................................{22} </strong></p>

    </body>
</html>
"""

for i in ows_dict:
    final_html = main_html.format(
        ows_dict[i]['First'],
        ows_dict[i]['Last'],
        ows_dict[i]['ID Number'],
        ows_dict[i]['Date Started'],
        ows_dict[i]['Rate Per Day'],
        ows_dict[i]['Rate Per Month'],
        ows_dict[i]['Total Work Hours'],
        ows_dict[i]['Total Overtime Hours'],
        ows_dict[i]['Total Late Hours'],
        ows_dict[i]['Total Basic Salary'],
        ows_dict[i]['Total Overtime Pay'],
        ows_dict[i]['Total Late Deduction'],
        ows_dict[i]['Total Gross Salary'],
        ows_dict[i]['Total Debt'],
        ows_dict[i]['Total Deductibles'],
        ows_dict[i]['Total Bonus'],
        ows_dict[i]['Total Standing Salary'],
        ows_dict[i]['Total Income Tax'],
        ows_dict[i]['Total Philhealth'],
        ows_dict[i]['Total Pagibig'],
        ows_dict[i]['Total SSS Employee Contribution'],
        ows_dict[i]['Total SSS Employer Contribution'],
        ows_dict[i]['Salary for the Week'])
    first_name = ows_dict[i]['First']
    last_name = ows_dict[i]['Last']
    pay_slip_name = directory + "/" + folder_name + "/" + "html-" + folder_name + "/" + "Pay Slip for"+ " " + first_name + " "+ last_name + ".html"
    with open(pay_slip_name, "w") as file:
        file.write(final_html)
